refactor(health-dashboard): split 757-LOC monolith into 12 focused files
Behavior preserved 1:1. Structure mirrors app-shell/ convention.

- components/health/types.ts — Point, MetricKey, HeroCard, ChartConfig, Props
- components/health/tones.ts — per-metric color palette (TONES)
- components/health/helpers.ts — groupDailyAverage, summarizeWindow,
  formatDateLabel, buildLinePath, getStats, getDelta
- components/health/empty-state.tsx + stat-pill.tsx — shared primitives
- components/health/trend-svg.tsx — SVG chart renderer
- components/health/trend-card.tsx — card wrapper around TrendSvg
- components/health/chart-modal.tsx — expanded-view modal
- components/health/hero-vitals.tsx — "Vitales del día" gradient band
- components/health/sleep-section.tsx — sleep analysis + 7d/bucket panel
- components/health/workouts-section.tsx — workouts + mix panel
- components/health/health-dashboard-view.tsx — orchestrator (useState +
  useMemo derivations, hero + chart configs, composition)
- components/health/health-range-selector.tsx — moved from top-level
  (sole importer was app/health/page.tsx)

No changes to UI, styles, API calls, hook semantics, or exported symbol
names. app/health/page.tsx updated to new import paths.

Verified: npm run typecheck clean, npx eslint components/health/
app/health/page.tsx clean. Build prerender fails only due to missing
Supabase env in this worktree (compile + TS check succeed).
